Extracting a superclass of InternalDistributedMember
We need a concrete MemberIdentifier implementation in geode-membership
to test membership.

Extracting a superclass of InternalDistributedMember,
MemberIdentifierImpl, which contains pretty much all of the
InternalDistributedMember logic. InternalDistributedMember now just
extends that class and implements DistributedMember.
