+++ geode-core/src/integrationTest/java/org/apache/geode/distributed/internal/membership/gms/MemberIdentifierFactoryImplTest.java
@@ -24,10 +24,11 @@ import java.util.Arrays;
 import org.junit.Before;
 import org.junit.Test;
 
-import org.apache.geode.distributed.internal.membership.InternalDistributedMember;
 import org.apache.geode.distributed.internal.membership.gms.api.MemberData;
 import org.apache.geode.distributed.internal.membership.gms.api.MemberDataBuilder;
 import org.apache.geode.distributed.internal.membership.gms.api.MemberIdentifier;
+import org.apache.geode.distributed.internal.membership.gms.api.MemberIdentifierFactoryImpl;
+import org.apache.geode.distributed.internal.membership.gms.api.MemberIdentifierImpl;
 import org.apache.geode.internal.inet.LocalHostUtil;
 import org.apache.geode.internal.serialization.Version;
 
@@ -112,10 +113,10 @@ public class MemberIdentifierFactoryImplTest {
   public void membersAreEqual() {
     MemberData memberData = MemberDataBuilder.newBuilderForLocalHost("hostname")
         .setMembershipPort(1).build();
-    InternalDistributedMember member1 = factory.create(memberData);
+    MemberIdentifierImpl member1 = factory.create(memberData);
     memberData = MemberDataBuilder.newBuilderForLocalHost("hostname")
         .setMembershipPort(1).build();
-    InternalDistributedMember member2 = factory.create(memberData);
+    MemberIdentifierImpl member2 = factory.create(memberData);
     assertThat(factory.getComparator().compare(member1, member2)).isZero();
   }
 
@@ -123,10 +124,10 @@ public class MemberIdentifierFactoryImplTest {
   public void membersAreNotEqual() {
     MemberData memberData = MemberDataBuilder.newBuilderForLocalHost("hostname")
         .setMembershipPort(1).build();
-    InternalDistributedMember member1 = factory.create(memberData);
+    MemberIdentifierImpl member1 = factory.create(memberData);
     memberData = MemberDataBuilder.newBuilderForLocalHost("hostname")
         .setMembershipPort(2).build();
-    InternalDistributedMember member2 = factory.create(memberData);
+    MemberIdentifierImpl member2 = factory.create(memberData);
     assertThat(factory.getComparator().compare(member1, member2)).isLessThan(0);
     assertThat(factory.getComparator().compare(member2, member1)).isGreaterThan(0);
   }
