+++ geode-core/src/main/java/org/apache/geode/internal/cache/TXCommitMessage.java
@@ -74,6 +74,7 @@ import org.apache.geode.internal.cache.versions.VersionTag;
 import org.apache.geode.internal.offheap.annotations.Released;
 import org.apache.geode.internal.serialization.DeserializationContext;
 import org.apache.geode.internal.serialization.SerializationContext;
+import org.apache.geode.internal.serialization.StaticSerialization;
 import org.apache.geode.internal.serialization.Version;
 import org.apache.geode.logging.internal.executors.LoggingThread;
 import org.apache.geode.logging.internal.log4j.api.LogService;
@@ -930,11 +931,11 @@ public class TXCommitMessage extends PooledDistributionMessage
   }
 
   private boolean hasFlagsField(final DataOutput out) {
-    return hasFlagsField(InternalDataSerializer.getVersionForDataStream(out));
+    return hasFlagsField(StaticSerialization.getVersionForDataStream(out));
   }
 
   private boolean hasFlagsField(final DataInput in) {
-    return hasFlagsField(InternalDataSerializer.getVersionForDataStream(in));
+    return hasFlagsField(StaticSerialization.getVersionForDataStream(in));
   }
 
   private boolean hasFlagsField(final Version version) {
@@ -1520,7 +1521,7 @@ public class TXCommitMessage extends PooledDistributionMessage
         this.preserializedBuffer.rewind();
         this.preserializedBuffer.sendTo(out);
       } else if (this.refCount > 1) {
-        Version v = InternalDataSerializer.getVersionForDataStream(out);
+        Version v = StaticSerialization.getVersionForDataStream(out);
         HeapDataOutputStream hdos = new HeapDataOutputStream(1024, v);
         basicToData(hdos, context, useShadowKey);
         this.preserializedBuffer = hdos;
