+++ geode-core/src/main/java/org/apache/geode/internal/cache/TXRegionLockRequestImpl.java
@@ -29,6 +29,7 @@ import org.apache.geode.cache.CacheClosedException;
 import org.apache.geode.internal.InternalDataSerializer;
 import org.apache.geode.internal.cache.locks.TXRegionLockRequest;
 import org.apache.geode.internal.logging.log4j.LogMarker;
+import org.apache.geode.internal.serialization.StaticSerialization;
 import org.apache.geode.internal.serialization.Version;
 import org.apache.geode.logging.internal.log4j.api.LogService;
 
@@ -118,7 +119,7 @@ public class TXRegionLockRequestImpl implements TXRegionLockRequest {
       if (cache != null && size > 0) {
         this.r = (LocalRegion) cache.getRegion(this.regionPath);
       }
-      if (InternalDataSerializer.getVersionForDataStream(in).compareTo(Version.GEODE_1_10_0) >= 0) {
+      if (StaticSerialization.getVersionForDataStream(in).compareTo(Version.GEODE_1_10_0) >= 0) {
         this.entryKeys = readEntryKeyMap(size, in);
       } else {
         this.entryKeys = readEntryKeySet(size, in);
@@ -182,7 +183,7 @@ public class TXRegionLockRequestImpl implements TXRegionLockRequest {
   @Override
   public void toData(DataOutput out) throws IOException {
     DataSerializer.writeString(getRegionFullPath(), out);
-    if (InternalDataSerializer.getVersionForDataStream(out).compareTo(Version.GEODE_1_10_0) >= 0) {
+    if (StaticSerialization.getVersionForDataStream(out).compareTo(Version.GEODE_1_10_0) >= 0) {
       InternalDataSerializer.writeHashMap(this.entryKeys, out);
     } else {
       HashSet hashset = new HashSet(this.entryKeys.keySet());
