+++ geode-core/src/main/java/org/apache/geode/internal/cache/partitioned/GetMessage.java
@@ -40,7 +40,6 @@ import org.apache.geode.distributed.internal.ReplyProcessor21;
 import org.apache.geode.distributed.internal.ReplySender;
 import org.apache.geode.distributed.internal.membership.InternalDistributedMember;
 import org.apache.geode.internal.Assert;
-import org.apache.geode.internal.InternalDataSerializer;
 import org.apache.geode.internal.cache.BucketRegion;
 import org.apache.geode.internal.cache.BucketRegion.RawValue;
 import org.apache.geode.internal.cache.CachedDeserializableFactory;
@@ -61,6 +60,7 @@ import org.apache.geode.internal.logging.log4j.LogMarker;
 import org.apache.geode.internal.offheap.OffHeapHelper;
 import org.apache.geode.internal.serialization.DeserializationContext;
 import org.apache.geode.internal.serialization.SerializationContext;
+import org.apache.geode.internal.serialization.StaticSerialization;
 import org.apache.geode.internal.serialization.Version;
 import org.apache.geode.internal.util.BlobHelper;
 import org.apache.geode.logging.internal.log4j.api.LogService;
@@ -461,7 +461,7 @@ public class GetMessage extends PartitionMessageWithDirectReply {
       this.valueType = flags;
       this.valueInBytes = DataSerializer.readByteArray(in);
       if (flags != VALUE_IS_BYTES) {
-        this.remoteVersion = InternalDataSerializer.getVersionForDataStreamOrNull(in);
+        this.remoteVersion = StaticSerialization.getVersionForDataStreamOrNull(in);
       }
       if (hasVersionTag) {
         this.versionTag = (VersionTag) DataSerializer.readObject(in);
