+++ geode-core/src/main/java/org/apache/geode/pdx/internal/PdxType.java
@@ -32,6 +32,7 @@ import org.apache.geode.DataSerializable;
 import org.apache.geode.DataSerializer;
 import org.apache.geode.internal.InternalDataSerializer;
 import org.apache.geode.internal.cache.tier.sockets.OldClientSupportService;
+import org.apache.geode.internal.serialization.StaticSerialization;
 import org.apache.geode.internal.serialization.Version;
 import org.apache.geode.pdx.PdxFieldAlreadyExistsException;
 import org.apache.geode.pdx.internal.AutoSerializableManager.AutoClassInfo;
@@ -133,7 +134,7 @@ public class PdxType implements DataSerializable {
       // to set noDomainClass to true.
       // For this reason the pdx delete-field command should only be used after
       // all member have been upgraded to 8.1 or later.
-      Version sourceVersion = InternalDataSerializer.getVersionForDataStream(out);
+      Version sourceVersion = StaticSerialization.getVersionForDataStream(out);
       if (sourceVersion.compareTo(Version.GFE_81) >= 0) {
         if (this.hasDeletedField) {
           bits |= HAS_DELETED_FIELD_BIT;
