+++ geode-serialization/src/main/java/org/apache/geode/internal/serialization/StaticSerialization.java
@@ -494,4 +494,64 @@ public class StaticSerialization {
             String.format("unexpected typeCode: %s", typeCode));
     }
   }
+
+  /**
+   * Get the {@link Version} of the peer or disk store that created this
+   * {@link DataInput}. Returns
+   * null if the version is same as this member's.
+   */
+  public static Version getVersionForDataStreamOrNull(DataInput in) {
+    // check if this is a versioned data input
+    if (in instanceof VersionedDataStream) {
+      return ((VersionedDataStream) in).getVersion();
+    } else {
+      // assume latest version
+      return null;
+    }
+  }
+
+  /**
+   * Get the {@link Version} of the peer or disk store that created this
+   * {@link DataInput}.
+   */
+  public static Version getVersionForDataStream(DataInput in) {
+    // check if this is a versioned data input
+    if (in instanceof VersionedDataStream) {
+      final Version v = ((VersionedDataStream) in).getVersion();
+      return v != null ? v : Version.CURRENT;
+    } else {
+      // assume latest version
+      return Version.CURRENT;
+    }
+  }
+
+  /**
+   * Get the {@link Version} of the peer or disk store that created this
+   * {@link DataOutput}.
+   */
+  public static Version getVersionForDataStream(DataOutput out) {
+    // check if this is a versioned data output
+    if (out instanceof VersionedDataStream) {
+      final Version v = ((VersionedDataStream) out).getVersion();
+      return v != null ? v : Version.CURRENT;
+    } else {
+      // assume latest version
+      return Version.CURRENT;
+    }
+  }
+
+  /**
+   * Get the {@link Version} of the peer or disk store that created this
+   * {@link DataOutput}. Returns
+   * null if the version is same as this member's.
+   */
+  public static Version getVersionForDataStreamOrNull(DataOutput out) {
+    // check if this is a versioned data output
+    if (out instanceof VersionedDataStream) {
+      return ((VersionedDataStream) out).getVersion();
+    } else {
+      // assume latest version
+      return null;
+    }
+  }
 }
